added firebase deployment
