+++ .firebaserc
@@ -1,5 +1,5 @@
 {
-  "projects": {
-    "default": "portfolio-43110"
-  }
+	"projects": {
+		"default": "portfolio-43110"
+	}
 }
