+++ .gitignore
@@ -24,6 +24,7 @@ npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
 .pnpm-debug.log*
+ui-debug.log
 
 # local env files
 .env*.local
