+++ firebase.json
@@ -1,6 +1,16 @@
 {
-  "hosting": {
-    "public": "out",
-    "ignore": ["firebase.json", "**/.*", "**/node_modules/**"]
-  }
+	"hosting": {
+		"public": "public",
+		"ignore": ["firebase.json", "**/.*", "**/node_modules/**"],
+		"rewrites": [
+			{
+				"source": "**",
+				"function": "nextjsFunc"
+			}
+		]
+	},
+	"functions": {
+		"source": ".",
+		"runtime": "nodejs14"
+	}
 }
