+++ package.json
@@ -1,36 +1,39 @@
 {
-  "name": "next-portfolio",
-  "version": "0.1.0",
-  "private": true,
-  "scripts": {
-    "dev": "next dev",
-    "build": "next build && next export",
-    "start": "next start",
-    "lint": "next lint",
-    "pretty": "prettier --write \"./**/*.{js,jsx,json}\""
-  },
-  "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "^6.1.1",
-    "@fortawesome/free-solid-svg-icons": "^6.1.1",
-    "bulma": "^0.9.4",
-    "bulma-prefers-dark": "^0.1.0-beta.1",
-    "next": "12.1.6",
-    "prettier": "^2.6.2",
-    "react": "18.1.0",
-    "react-bulma-components": "^4.1.0",
-    "react-dom": "18.1.0",
-    "sass": "^1.52.2"
-  },
-  "devDependencies": {
-    "@types/node": "17.0.39",
-    "@types/react": "18.0.11",
-    "@types/react-dom": "18.0.5",
-    "@typescript-eslint/eslint-plugin": "^5.27.1",
-    "@typescript-eslint/parser": "^5.27.1",
-    "autoprefixer": "^10.4.7",
-    "eslint": "^8.17.0",
-    "eslint-config-next": "12.1.6",
-    "eslint-plugin-react": "^7.30.0",
-    "typescript": "4.7.3"
-  }
+	"private": true,
+	"main": "firebaseFunctions.js",
+	"scripts": {
+		"dev": "next src/",
+		"build": "next build src/",
+		"start": "next start src/",
+		"serve": "npm run build && firebase emulators:start --only functions,hosting",
+		"shell": "npm run build && firebase functions:shell",
+		"deploy": "firebase deploy --only functions,hosting",
+		"logs": "firebase functions:log"
+	},
+	"dependencies": {
+		"@fortawesome/fontawesome-svg-core": "^6.1.1",
+		"@fortawesome/free-solid-svg-icons": "^6.1.1",
+		"bulma": "^0.9.4",
+		"bulma-prefers-dark": "^0.1.0-beta.1",
+		"firebase-admin": "^10.2.0",
+		"firebase-functions": "^3.21.2",
+		"next": "12.1.6",
+		"prettier": "^2.6.2",
+		"react": "18.1.0",
+		"react-bulma-components": "^4.1.0",
+		"react-dom": "18.1.0",
+		"sass": "^1.52.2"
+	},
+	"devDependencies": {
+		"@types/node": "17.0.39",
+		"@types/react": "18.0.11",
+		"@types/react-dom": "18.0.5",
+		"@typescript-eslint/eslint-plugin": "^5.27.1",
+		"@typescript-eslint/parser": "^5.27.1",
+		"autoprefixer": "^10.4.7",
+		"eslint": "^8.17.0",
+		"eslint-config-next": "12.1.6",
+		"eslint-plugin-react": "^7.30.0",
+		"typescript": "4.7.3"
+	}
 }
