+++ src/styles/globals.scss
@@ -1,0 +1,73 @@
+@use 'sass:color';
+@import './palette.scss';
+
+html,
+body {
+	padding: 0;
+	margin: 0;
+	font-family: -apple-system, BlinkMacSystemFont, Segoe UI, Roboto, Oxygen,
+		Ubuntu, Cantarell, Fira Sans, Droid Sans, Helvetica Neue, sans-serif;
+	background-color: $primary;
+}
+
+h1,
+h2,
+h3,
+h4,
+h5,
+h6,
+p,
+li,
+.title {
+	color: $light-grey;
+}
+
+.button {
+	background-color: $secondary;
+}
+
+ul {
+	list-style: square;
+}
+
+a {
+	color: inherit;
+	text-decoration: none;
+}
+
+* {
+	box-sizing: border-box;
+}
+
+#root > * {
+	max-width: 70%;
+	min-height: 100vh;
+	margin: 0 auto;
+	text-align: left;
+	display: flex;
+	justify-content: center;
+	align-items: center;
+	flex-direction: column;
+}
+
+.title.is-1,
+.title.is-2 {
+	text-align: center;
+}
+
+.title.is-1 {
+	font-size: 2.5vw;
+	margin-bottom: 0.5rem;
+}
+
+.title.is-2 {
+	font-size: 2.3vw;
+	margin-bottom: 0;
+}
+
+.is-256x256 {
+	width: 256px;
+	height: 256px;
+}
+
+@import '../../node_modules/bulma/sass/base/generic.sass';
